Added meshzoo to requirements, refactored mesh_builders
